updated crud process to cloud

+++ Models/DeleteBook.cs
@@ -1,5 +1,6 @@
 using System.Data.SQLite;
 using API.Models.Interfaces;
+using MySql.Data.MySqlClient;
 
 namespace API.Models
 {
@@ -7,16 +8,25 @@ namespace API.Models
     {
         public void RemoveBook(int id)
         {
-            string cs = @"URI=file:C:\Users\jslucas\source\repos\mis321\database\book.db";
-            using var con = new SQLiteConnection(cs);
-            con.Open();
+            // string cs = @"URI=file:C:\Users\jslucas\source\repos\mis321\database\book.db";
+            // using var con = new SQLiteConnection(cs);
+            // con.Open();
 
-            using var cmd = new SQLiteCommand(con);
+            // using var cmd = new SQLiteCommand(con);
+            DBConnect db = new DBConnect();
+            bool isOpen = db.OpenConnection();
 
-            cmd.CommandText = @"DELETE FROM books WHERE id=@id";
-            cmd.Parameters.AddWithValue("@id",id);
-            cmd.Prepare();
-            cmd.ExecuteNonQuery();
+            if(isOpen){
+                MySqlConnection conn = db.GetConn();
+                MySqlCommand cmd = new MySqlCommand();
+
+                cmd.Connection = conn;
+
+                cmd.CommandText = @"DELETE FROM books WHERE id=@id";
+                cmd.Parameters.AddWithValue("@id",id);
+                cmd.Prepare();
+                cmd.ExecuteNonQuery();
+            }
         }
     }
 }
